requirements.txt
python-telegram-bot[job-queue,webhooks]
fastapi
uvicorn
gunicorn
langgraph
langchain
langchain-text-splitters
langchain-openai
openai
mcp
qdrant-client
python-dotenv
pydantic
pydantic-settings
httpx
tiktoken
beautifulsoup4
pypdf
matplotlib
